Refactor cyral_repository_access_rules
