+++ cyral/internal/repository/accessrules/constants.go
@@ -1,0 +1,5 @@
+package accessrules
+
+const (
+	resourceName = "cyral_repository_access_rules"
+)
